Assignment 3/requirements.txt
absl-py==0.11.0
alembic==1.5.4
altair==4.1.0
apache-airflow==2.0.1
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.1
apispec==3.3.2
appnope==0.1.2
argcomplete==1.12.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
boto3==1.17.10
botocore==1.20.10
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
cattrs==1.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
cmake==3.18.4.post1
colorama==0.4.4
colorlog==4.7.2
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
#dlib==19.21.1
dnspython==2.1.0
docutils==0.16
email-validator==1.1.2
entrypoints==0.3
fastapi==0.63.0
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.13
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
gunicorn==19.10.0
h11==0.12.0
h5py==2.10.0
html5lib==1.1
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
install==1.3.4
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
lockfile==0.12.2
lxml==4.6.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.4
mistune==0.8.4
natsort==7.1.1
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
openapi-spec-validator==0.2.9
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prison==0.1.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8
pydeck==0.6.1
Pygments==2.8.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rich==9.2.0
rsa==4.7.1
s3transfer==0.3.4
scipy==1.6.0
Send2Trash==1.5.0
setproctitle==1.2.2
six==1.15.0
smmap==3.0.5
soupsieve==2.2
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
streamlit==0.76.0
svgwrite==1.4.1
swagger-ui-bundle==0.0.8
tabulate==0.8.8
tenacity==6.2.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.8.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.26.3
uvicorn==0.13.4
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.0
